Update Rust edition + restructure tests
