+++ Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "ohsl"
-version = "0.7.6"
+version = "0.7.7"
 authors = ["anthonyoneill <ajoneill@protonmail.com>"]
-edition = "2018"
+edition = "2021"
 description = "A collection of numerical routines and mathematical types for use in scientific computing."
 license = "MIT"
 repository = "https://github.com/anthonyoneill/ohsl"
