+++ tests/matrix/matrix.rs
@@ -1,0 +1,14 @@
+mod construction;
+mod arithmetic;
+mod operations;
+mod functions;
+mod solve;
+
+/*#[test]
+fn file_output() {
+    let mut m = Matrix::<i32>::new( 2, 2, 1 );
+    m[0][1] = 2;
+    m[1][0] = 3; 
+    m[1][1] = 4;
+    m.output( "./output.txt" );
+}*/
