+++ tests/newton/newton.rs
@@ -3,7 +3,7 @@ use ohsl::matrix::Mat64;
 use ohsl::newton::Newton;
 
 #[test]
-fn test_newton_constructor() {
+fn constructor() {
     let newton = Newton::<f64>::new( 0.0 );
     let parameters = newton.parameters();
     assert_eq!( parameters.0, 1.0e-8 );
@@ -13,7 +13,7 @@ fn test_newton_constructor() {
 }
 
 #[test]
-fn test_newton_parameters() {
+fn parameters() {
     let mut newton = Newton::<f64>::new( 0.0 );
     newton.tolerance( 1.0e-6 );
     newton.delta( 1.0e-7 );
@@ -26,17 +26,17 @@ fn test_newton_parameters() {
     assert_eq!( parameters.3, 1.0 );
 }
 
-fn myfunction(x: f64) -> f64 {
+fn x_squared_minus_four(x: f64) -> f64 {
     x * x - 4.0
 }
 
 #[test]
-fn test_newton_solve_f64() {
+fn solve_f64() {
     let mut newton = Newton::<f64>::new( 1.0 );
-    let solution = newton.solve( &myfunction );
+    let solution = newton.solve( &x_squared_minus_four );
     assert_eq!( solution.unwrap(), 2.0 );
     newton.guess( -1.0 );
-    let solution = newton.solve( &myfunction ).unwrap();
+    let solution = newton.solve( &x_squared_minus_four ).unwrap();
     assert_eq!( solution, -2.0 );
 }
 
@@ -53,7 +53,7 @@ fn vecfunc( x: Vec64 ) -> Vec64 {
 }
 
 #[test]
-fn test_newton_jacobian() {
+fn jacobian() {
     let point = Vec64::create( vec![ 1.0, 1.0 ] );
     let jacobian = Mat64::jacobian( point, &vecfunc, 1.0e-8 );
     assert_eq!( jacobian.rows(), 2 );
@@ -65,7 +65,7 @@ fn test_newton_jacobian() {
 }
 
 #[test]
-fn test_newton_solve_vec64() {
+fn solve_vec64() {
     let guess = Vec64::create( vec![ 0.5, 0.25 ] );
     let newton = Newton::<Vec64>::new( guess );
     let solution = newton.solve( &vecfunc ).unwrap();
@@ -87,7 +87,7 @@ fn jacfunc( x: Vec64 ) -> Mat64 {
 }
 
 #[test]
-fn test_newton_exact_jacobian_solve() {
+fn exact_jacobian_solve() {
     let guess = Vec64::create( vec![ 0.5, 0.25 ] );
     let newton = Newton::<Vec64>::new( guess );
     let solution = newton.solve_jacobian( &vecfunc, &jacfunc ).unwrap();
