+++ tests/tests.rs
@@ -1,0 +1,20 @@
+#[path = "complex/complex.rs"]
+mod complex;
+
+#[path = "vector/vector.rs"]
+mod vector;
+
+#[path = "matrix/matrix.rs"]
+mod matrix;
+
+#[path = "mesh1d/mesh1d.rs"]
+mod mesh1d;
+
+#[path = "mesh2d/mesh2d.rs"]
+mod mesh2d;
+
+#[path = "newton/newton.rs"]
+mod newton;
+
+#[path = "sparse/sparse.rs"]
+mod sparse;
