+++ tests/vector/vector.rs
@@ -1,0 +1,15 @@
+//use ohsl::vector::{Vector, Vec64};
+
+mod construction;
+mod arithmetic;
+mod basic_functions;
+mod sequence;
+mod norms;
+mod random;
+mod sort;
+
+/*#[test]
+fn file_output() {
+    let u = Vector::<u32>::create( vec![ 1, 2, 3, 4, 5 ] );
+    u.output( "./output.txt" );
+}*/
